requirements.txt
torch>=1.12.0
numpy>=1.21.0
Pillow>=8.0.0